refactor(): rename vue component pending-pr -> open-pr
