+++ app/javascript/components/open-pr.vue
