+++ app/views/github_users/_profile_card.html.erb
@@ -41,9 +41,9 @@
     </div>
   <div class="profile-prs">
     <% @open_prs.each do |pr| %>
-        <pending-pr
+        <open-pr
           pr="<%= pr.title %>"
-        ></pending-pr>
+        ></open-pr>
     <% end %>
   </div>
   <div class="profile__section">
